requirements_simple.txt
# Core Backend Dependencies
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.32
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
scikit-learn==1.5.1

# WebSocket and Real-time Communication
websockets==12.0
python-socketio>=5.8.0

# Basic Sentiment Analysis Dependencies
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
beautifulsoup4==4.12.2
aiohttp==3.9.5
redis==5.0.1

# VectorBT backtesting dependencies
vectorbt>=0.25.0
numba>=0.56.0

# Basic LLM Dependencies (minimal set)
openai>=1.0.0
anthropic>=0.21.0
tiktoken>=0.5.0

# Production Database Dependencies
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
alembic>=1.13.0

# Configuration Management
pyyaml>=6.0.1

# Authentication and Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
cryptography>=41.0.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0